Single_Inheritence_Method_OverLoading
This file contains the basic implementation of inheritance with method overloading. It will be helpful to understand how methods can be overloaded within a class hierarchy.

Method overloading refers to the process of defining multiple methods with the same name but different parameter lists. This file includes code that is clear and simple, making it easier to understand the logic of method overloading in inheritance. The code has been executed and verified without any errors.